[refactor] 도메인의 성격에 따라 각 패키지로 분리
