+++ src/main/java/nextstep/ladder/domain/game/GameResults.java
@@ -1,4 +1,6 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.game;
+
+import nextstep.ladder.domain.game.GameResult;
 
 import java.util.ArrayList;
 import java.util.Collections;
