+++ src/main/java/nextstep/ladder/domain/point/PointState.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.point;
 
 import nextstep.UniformRandomBooleanProvider;
