+++ src/main/java/nextstep/ladder/domain/point/Points.java
@@ -1,4 +1,6 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.point;
+
+import nextstep.ladder.domain.point.Point;
 
 import java.util.ArrayList;
 import java.util.Collections;
