+++ src/test/java/nextstep/ladder/LineTest.java
@@ -1,6 +1,6 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.Lines;
+import nextstep.ladder.domain.line.Lines;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
