+++ src/test/java/nextstep/ladder/ParticipantTest.java
@@ -1,6 +1,6 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.Participant;
+import nextstep.ladder.domain.participant.Participant;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
