+++ src/test/java/nextstep/ladder/ParticipantsTest.java
@@ -1,7 +1,7 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.Participant;
-import nextstep.ladder.domain.Participants;
+import nextstep.ladder.domain.participant.Participant;
+import nextstep.ladder.domain.participant.Participants;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
