+++ src/test/java/nextstep/ladder/PointStateTest.java
@@ -1,6 +1,6 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.PointState;
+import nextstep.ladder.domain.point.PointState;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
