+++ src/test/java/nextstep/ladder/PointTest.java
@@ -1,7 +1,7 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.Point;
-import nextstep.ladder.domain.PointState;
+import nextstep.ladder.domain.point.Point;
+import nextstep.ladder.domain.point.PointState;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
